Sửa gọi dịch vụ cùng booking

+++ database/migrations/2025_06_25_145518_update_booking_service_add_room_id.php
@@ -6,18 +6,11 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
         Schema::table('booking_service', function (Blueprint $table) {
             if (!Schema::hasColumn('booking_service', 'room_id')) {
-                $table->foreignId('room_id')
-                    ->after('booking_id')
-                    ->nullable()
-                    ->constrained('rooms')
-                    ->cascadeOnDelete();
+                $table->foreignId('room_id')->after('booking_id')->nullable()->constrained('rooms')->nullOnDelete();
             }
         });
     }
@@ -28,9 +21,12 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('booking_service', function (Blueprint $table) {
-            // Phải drop foreign key trước
-            $table->dropForeign(['room_id']);
-            $table->dropColumn('room_id');
+            if (Schema::hasColumn('booking_service', 'room_id')) {
+                $table->dropForeign(['room_id']);
+                $table->dropColumn('room_id');
+            }
         });
     }
 };
+
+
